Add sales income script in python, start on electron python build
The sales income script is the basic script I need, so really the extra
stuff is just for fun. I want ot try and make an electron python build,
I tried to do the original script in javascript but everything is like
using a rusty spoon compared to pandas, so I think this is a better
plan.